# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Цифровой двойник цеха — точка входа",
    "version": "1.0.0",
    "description": "Единая документированная точка входа двойника. По этим адресам ходят имитаторы 1С в прототипе и пойдут настоящие системы завода.\n\n- Все события проверяются схемой; ошибки сохраняются и видны на экране «Источники данных».\n- Повтор события с тем же `eventId` не создаёт дубль (идемпотентность).\n- Время — ISO 8601 с часовым поясом, завод работает по UTC+5.\n- Контроллеры и видеоаналитика подключаются по MQTT, см. «Документация MQTT (AsyncAPI)».\n\nТипичная интеграция с 1С:MES/QLS/WMS — HTTP-сервисы или OData; конкретная выгрузка настраивается со специалистами завода."
  },
  "servers": [
    {
      "url": "/",
      "description": "Шлюз двойника"
    }
  ],
...
  "paths": {
...
    "/api/v1/plant/config": {
      "get": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Текущая применённая конфигурация (с живым состоянием связи)",
        "responses": {
          "200": {
            "description": "Конфигурация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantConfig"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Применить новую версию",
        "description": "Проверяет и применяет. Версию, время и автора ставит шлюз. Нельзя удалить оборудование, по которому открыт инцидент.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantConfigInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Применено: новая версия и список изменений",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plant/config/validate": {
      "post": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Проверка без применения",
        "description": "Ошибки (нельзя применить) и предупреждения — по-русски, с путём к участку, станции или оборудованию.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantConfigInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог проверки",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantValidation"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plant/config/preview-impact": {
      "post": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Влияние изменений без применения",
        "description": "Изменения человеческим языком, пропускная способность участков до и после, узкое место.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantConfigInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Влияние",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantImpact"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plant/config/rollback/{ver}": {
      "post": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Откат к версии",
        "description": "Создаёт новую версию — копию указанной. История остаётся линейной.",
        "parameters": [
          {
            "name": "ver",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Откат применён",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Версии нет в истории"
          }
        }
      }
    },
...
    "/api/v1/plant/config/import": {
      "post": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Загрузить конфигурацию из JSON и применить",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantConfigInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Применено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/equipment/{id}/connection/test": {
      "post": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Проверка подключения оборудования",
        "description": "Имитатор (демо) опрашивается по-настоящему и возвращает живые значения. Для OPC UA, Modbus, S7 и SCADA — проверка адреса и попытка TCP-подключения с таймаутом; успешное подключение к несуществующему устройству не изображается.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EquipmentCode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantConnectionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог проверки",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResult"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Оборудования нет в применённой конфигурации"
          }
        }
      }
    },
    "/api/v1/equipment/{id}/connection": {
      "put": {
        "tags": [
          "Конфигурация завода"
        ],
        "summary": "Сохранить подключение оборудования",
        "description": "Создаёт новую версию конфигурации. Статус «на связи» появится, когда по оборудованию реально придут данные.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EquipmentCode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantConnectionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Сохранено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Оборудования нет в применённой конфигурации"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelId": {
        "type": "string",
        "enum": [
          "onix",
          "cobalt",
          "j7"
        ],
        "description": "Модель: onix — Chevrolet Onix, cobalt — Chevrolet Cobalt, j7 — JAC J7"
      },
...
      "Timestamp": {
        "type": "string",
        "format": "date-time",
        "description": "Время в ISO 8601 с часовым поясом",
        "examples": [
          "2026-10-07T13:40:00+05:00"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
...
      "EquipmentCode": {
        "type": "string",
        "pattern": "^[A-Z0-9][A-Z0-9_-]{0,31}$",
        "description": "Код оборудования, уникальный на заводе. Он же — сегмент MQTT-топика",
        "examples": [
          "BOOTH-02"
        ]
      },
...
      "StageCode": {
        "type": "string",
        "pattern": "^[a-z][a-z0-9-]{0,31}$",
        "description": "Код участка. Он же — поле area в событиях и сегмент MQTT-топика",
        "examples": [
          "paint"
        ]
      },
...
      "PlantColor": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "hex": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "finish": {
            "type": "string",
            "enum": [
              "solid",
              "metallic",
              "pearl"
            ],
            "description": "solid — без металлика, metallic — металлик, pearl — перламутр"
          },
          "models": {
            "description": "Для каких моделей доступен; нет — для любых",
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelId"
            }
          }
        },
        "required": [
          "code",
          "name",
          "hex",
          "finish"
        ],
        "description": "Цвет кузова из справочника 1С: код из производственного заказа → название и оттенок"
      },
...
      "PlantStage": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/StageCode"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "short": {
            "description": "Короткое название для подписей: «ОТК», «Склад ГП»",
            "type": "string",
            "maxLength": 20
          },
          "kind": {
            "type": "string",
            "enum": [
              "warehouse_in",
              "welding",
              "painting",
              "assembly",
              "inspection",
              "warehouse_out",
              "custom"
            ],
            "description": "Вид участка: склад комплектующих, сварка, окраска, сборка, контроль, склад готовой продукции, другой"
          },
          "order": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "models": {
            "description": "Какие модели обрабатывает: сварочные линии привязаны к моделям, окраска и сборка универсальные. Нет поля — любые модели",
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelId"
            }
          },
          "idPoints": {
            "description": "Точки отметки кузова на входе и выходе участка",
            "$ref": "#/components/schemas/PlantIdPoints"
          },
          "bufferAfter": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "capacity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 200
                  }
                },
                "required": [
                  "capacity"
                ]
              },
              {
                "type": "null"
              }
            ],
            "description": "Буфер до следующего участка, кузовов"
          },
          "inlet": {
            "description": "Общее оборудование на входе участка: все кузова проходят его до станций (ванна подготовки)",
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantEquipment"
            }
          },
          "stations": {
            "maxItems": 40,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantStation"
            }
          },
          "outlet": {
            "description": "Общее оборудование на выходе участка: все кузова проходят его после станций (сушильная печь)",
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantEquipment"
            }
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "order",
          "bufferAfter",
          "stations"
        ],
        "description": "Участок завода"
      },
      "PlantConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedBy": {
            "type": "string",
            "maxLength": 100
          },
          "comment": {
            "description": "Что изменилось в этой версии",
            "type": "string",
            "maxLength": 300
          },
          "stages": {
            "minItems": 2,
            "maxItems": 30,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantStage"
            }
          },
          "colors": {
            "description": "Справочник цветов кузова; нет — справочник по умолчанию",
            "maxItems": 40,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantColor"
            }
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "updatedAt",
          "updatedBy",
          "stages"
        ],
        "description": "Конфигурация завода: участки по потоку слева направо, их станции и оборудование"
      },
      "PlantConfigInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedBy": {
            "type": "string",
            "maxLength": 100
          },
          "comment": {
            "description": "Что изменилось в этой версии",
            "type": "string",
            "maxLength": 300
          },
          "stages": {
            "minItems": 2,
            "maxItems": 30,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantStage"
            }
          },
          "colors": {
            "description": "Справочник цветов кузова; нет — справочник по умолчанию",
            "maxItems": 40,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantColor"
            }
          }
        },
        "required": [
          "id",
          "name",
          "stages"
        ],
        "description": "Конфигурация завода для проверки или применения (версию и время ставит шлюз)"
      },
      "PlantConnectionInput": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "opcua",
              "modbus_tcp",
              "s7",
              "scada",
              "retrofit_sensor",
              "manual",
              "simulator",
              "none"
            ],
            "description": "opcua — OPC UA, modbus_tcp — Modbus TCP, s7 — Siemens S7, scada — система управления участком, retrofit_sensor — внешний датчик, manual — ручной ввод мастером, simulator — имитатор (демо), none — не подключено"
          },
          "endpoint": {
            "description": "Адрес: opc.tcp://10.20.1.15:4840, 10.20.1.30 (Modbus, S7), идентификатор датчика",
            "type": "string",
            "maxLength": 300
          },
          "params": {
            "description": "Параметры способа: режим безопасности OPC UA, порт и unit id Modbus, rack/slot S7",
            "type": "object",
            "propertyNames": {
              "type": "string",
              "maxLength": 40
            },
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          },
          "tagMap": {
            "description": "Какой тег источника что означает для двойника",
            "type": "object",
            "propertyNames": {
              "type": "string",
              "enum": [
                "state",
                "errorCode",
                "cycleCounter",
                "filterDpPa",
                "motorCurrentA",
                "vibrationMmS",
                "temperatureC",
                "pressureBar",
                "torqueNm"
              ]
            },
            "additionalProperties": {
              "type": "string",
              "maxLength": 200
            }
          }
        },
        "required": [
          "method"
        ],
        "description": "Подключение оборудования: способ, адрес, параметры и сопоставление тегов"
      },
      "PlantIssue": {
        "type": "object",
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "error",
              "warning"
            ]
          },
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stageId": {
            "type": "string"
          },
          "stationId": {
            "type": "string"
          },
          "equipmentId": {
            "type": "string"
          },
          "incidentId": {
            "type": "string"
          }
        },
        "required": [
          "level",
          "path",
          "message"
        ],
        "description": "Ошибка или предупреждение проверки состава цеха — по-русски"
      },
      "PlantValidation": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantIssue"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlantIssue"
            }
          }
        },
        "required": [
          "ok",
          "errors",
          "warnings"
        ],
        "description": "Итог проверки: ok — ошибок нет, можно применять"
      },
...
      "StageCapacity": {
        "type": "object",
        "properties": {
          "stageId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "stations": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "nominalPerShift": {
            "type": "number",
            "description": "По норме цикла: сумма станций, но не больше общего оборудования"
          },
          "effectivePerShift": {
            "type": "number",
            "description": "С учётом простоев за 30 дней и перекраски"
          },
          "limitedBy": {
            "description": "Что ограничивает участок, если не сумма станций",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "stageId",
          "name",
          "stations",
          "nominalPerShift",
          "effectivePerShift",
          "limitedBy"
        ],
        "description": "Пропускная способность участка, кузовов в смену"
      },
      "PlantImpact": {
        "type": "object",
        "properties": {
          "changes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Изменения человеческим языком: «+1 камера окраски»"
          },
          "before": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageCapacity"
            }
          },
          "after": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageCapacity"
            }
          },
          "bottleneckBefore": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "stageId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "stageId",
                  "name"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "bottleneckAfter": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "stageId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "stageId",
                  "name"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "forecast": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "before": {
                    "type": "number"
                  },
                  "after": {
                    "type": "number"
                  },
                  "target": {
                    "type": "number"
                  },
                  "gapBefore": {
                    "type": "number"
                  },
                  "gapAfter": {
                    "type": "number"
                  }
                },
                "required": [
                  "before",
                  "after",
                  "target",
                  "gapBefore",
                  "gapAfter"
                ]
              },
              {
                "type": "null"
              }
            ],
            "description": "Прогноз плана месяца до и после (P50)"
          },
          "validation": {
            "$ref": "#/components/schemas/PlantValidation"
          }
        },
        "required": [
          "changes",
          "before",
          "after",
          "bottleneckBefore",
          "bottleneckAfter",
          "forecast",
          "validation"
        ],
        "description": "Влияние изменений состава цеха — без применения"
      },
      "ConnectionTestResult": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "description": "Что сделать, если не получилось",
            "type": "string"
          },
          "values": {
            "description": "Живые значения тегов, если подключение удалось",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "enum": [
                    "state",
                    "errorCode",
                    "cycleCounter",
                    "filterDpPa",
                    "motorCurrentA",
                    "vibrationMmS",
                    "temperatureC",
                    "pressureBar",
                    "torqueNm"
                  ]
                },
                "label": {
                  "type": "string"
                },
                "tag": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "value": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "label",
                "tag",
                "value"
              ]
            }
          }
        },
        "required": [
          "ok",
          "message"
        ],
        "description": "Итог проверки подключения"
      }
    },
...
  }
}